Python-Scripts/requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
dill==0.2.7.1
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
futures==3.1.1
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.18.3
get==0.0.0
google-auth==1.1.1
google-cloud-core==0.27.1
google-cloud-language==0.29.0
google-gax==0.15.15
googleapis-common-protos==1.5.2
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.4.0
httplib2==0.10.3
idna==2.6
jdcal==1.3
oauth2client==4.1.2
openpyxl==2.4.8
ply==3.8
post==0.0.0
protobuf==3.4.0
public==0.0.0
pyasn1==0.3.6
pyasn1-modules==0.1.4
PySocks==1.6.7
query-string==0.0.0
request==0.0.0
requests==2.18.4
rsa==3.4.2
selenium==3.4.3
setupfiles==0.0.50
six==1.11.0
urllib==1.21.1
urllib3==1.22
